- Making urls public in env. - Removed google analytics. - Moved image pointers from azure to local. - Moved to NextJS v13.
