+++ next-sitemap.config.js
@@ -1,6 +1,6 @@
 /** @type {import('next-sitemap').IConfig} */
 module.exports = {
-    siteUrl: process.env.SITE_URL || 'https://dbexpression.com',
+    siteUrl: process.env.NEXT_PUBLIC_SITE_URL || 'https://dbexpression.com',
     generateRobotsTxt: true,
     generateIndexSitemap: false,
     robotsTxtOptions: {
