+++ package.json
@@ -18,19 +18,20 @@
     "autoprefixer": "^10.4.7",
     "clsx": "^1.2.0",
     "focus-visible": "^5.2.0",
-    "next": "^12.2.3",
+    "next": "^13.5.6",
     "next-sitemap": "^3.1.29",
     "postcss-focus-visible": "^6.0.4",
     "prism-react-renderer": "^1.3.5",
     "prismjs": "^1.29.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "sharp": "^0.33.3",
     "swiper": "^8.4.5",
     "tailwindcss": "^3.1.4"
   },
   "devDependencies": {
     "eslint": "8.19.0",
-    "eslint-config-next": "12.2.3",
+    "eslint-config-next": "^13.5.6",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.11"
   }
