+++ src/components/Hero.jsx
@@ -1,9 +1,9 @@
-import Image from 'next/future/image'
 import { Button } from '@/components/Button'
 import { Container } from '@/components/Container'
 import { HeroBackground } from '@/components/HeroBackground'
 import { Fragment } from 'react'
 import Highlight, { defaultProps } from 'prism-react-renderer'
+import Image from 'next/image'
 
 import blurCyanImage from '@/images/blur-cyan.png'
 import blurIndigoImage from '@/images/blur-indigo.png'
